+++ src/Exception/MethodNotAllowedException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace D5WHUB\Extend\Router\Exception;
+
+class MethodNotAllowedException extends RuntimeException
+{
+}