+++ src/Exception/NotFoundException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace D5WHUB\Extend\Router\Exception;
+
+class NotFoundException extends RuntimeException
+{
+}