+++ src/Manager.php
@@ -8,6 +8,8 @@
 namespace D5WHUB\Extend\Router;
 
 use D5WHUB\Extend\Router\Cache\CacheInterface;
+use D5WHUB\Extend\Router\Exception\MethodNotAllowedException;
+use D5WHUB\Extend\Router\Exception\NotFoundException;
 use D5WHUB\Extend\Router\Exception\RuntimeException;
 use D5WHUB\Extend\Router\Exception\SyntaxException;
 use D5WHUB\Extend\Router\Manager\Constants;
@@ -110,6 +112,8 @@ class Manager extends Constants
      * @param string $httpMethod
      * @param string $uri
      * @return Context
+     * @throws MethodNotAllowedException
+     * @throws NotFoundException
      * @throws RuntimeException
      * @noinspection PhpDocMissingThrowsInspection PhpUnhandledExceptionInspection
      */